+++ src/stories/Card/Card.js
@@ -9,10 +9,10 @@ export const createCard = ({
   const card = document.createElement('div');
 
   const headerHtml = `
-    <div class='card__header background-primary-plus-2'>
-      <h2>${header}</h2>
-    </div>
-  `;
+  <div class='card__header background-primary-plus-2'>
+    <h2>${header}</h2>
+  </div>
+`;
 
   const coreClass = padded ? 'card-padded' : 'card';
   const condensedClass = condensed ? 'card--condensed' : '';
@@ -24,18 +24,18 @@ export const createCard = ({
 
   if (body) {
     card.innerHTML += `
-      <div class='card__body'>
-        <p>${body}</p>
-      </div>
-    `;
+  <div class='card__body'>
+    <p>${body}</p>
+  </div>
+`;
   }
 
   if (footer) {
     card.innerHTML += `
-      <div class='card__footer background-primary-plus-3'>
-        <p>${footer}</p>
-      </div>
-    `;
+  <div class='card__footer background-primary-plus-3'>
+    <p>${footer}</p>
+  </div>
+`;
   }
 
   return card;
